服务基本完成 但都不解析

+++ SmartHotel/SmartHotel/Classes/Tools/SHConstraint.h
@@ -12,7 +12,7 @@
 
 #import <UIKit/UIKit.h>
 
-typedef NS_ENUM(Byte, SHRoomServerType) {
+typedef NS_ENUM(NSInteger, SHRoomServerType) {
     
     SHRoomServerTypeCleanOver,
     SHRoomServerTypeClean,
